requirements.txt
numpy>=1.19.0
scipy>=1.7.0
scikit-learn>=0.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
flask
torch>=1.9.0
streamlit>=1.0.0